+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 - Support deletion prevention for `kubectl gs template app` command
 
+### Fixed
+
+- Listen only on localhost for callbacks
+
 ## [2.42.0] - 2023-10-06
 
 ### Fixed